pyproject.toml
[project]
name = "tuutrag-open"
version = "0.1.0"
description = "Teaching to Understand, Understanding to Teach: Retreival Augmented Generation for Requirements Traceability"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-exports>=1.2.0",
]
